Fully replace synchronous implemenation with async
