+++ tests/urlscan/test_async.py
@@ -6,7 +6,7 @@ import pytest
 from aioresponses import aioresponses
 from pytest_mock import MockFixture
 
-from ..context import urlscan_async
+from ..context import urlscan
 
 
 # Utility function to allow for mocking async function returns
@@ -19,10 +19,9 @@ async def test_submit_scan_request(test_urlscan_params: Dict[str, Any],
                                    submit_response: Dict[str, Any]) -> None:
     with aioresponses() as mocked:
         mocked.post(test_urlscan_params["submit_url"], status=200, body=json.dumps(submit_response))
-        async with urlscan_async.UrlScanAsync(
-            api_key=test_urlscan_params["api_key"],
-            data_dir=test_urlscan_params["data_dir"]) as urlscan:
-            actual: UUID = await urlscan.submit_scan_request("https://www.test.com")
+        async with urlscan.UrlScan(api_key=test_urlscan_params["api_key"],
+                                   data_dir=test_urlscan_params["data_dir"]) as url_scan:
+            actual: UUID = await url_scan.submit_scan_request("https://www.test.com")
             assert UUID(submit_response["uuid"]) == actual
 
 
@@ -31,10 +30,10 @@ async def test_fetch_result(mocker,
                             test_urlscan_params: Dict[str, Any],
                             success_result_response: Dict[str, Any]) -> None:
     # Mock download_screenshot function
-    mock_download_screenshot: MockFixture = mocker.patch("src.urlscanio.urlscan_async.UrlScanAsync.download_screenshot")
+    mock_download_screenshot: MockFixture = mocker.patch("src.urlscanio.urlscan.UrlScan.download_screenshot")
     mock_download_screenshot.return_value = await return_async(test_urlscan_params["screenshot"]["path"])
     # Mock download_dom function
-    mock_download_dom: MockFixture = mocker.patch("src.urlscanio.urlscan_async.UrlScanAsync.download_dom")
+    mock_download_dom: MockFixture = mocker.patch("src.urlscanio.urlscan.UrlScan.download_dom")
     mock_download_dom.return_value = await return_async(test_urlscan_params["dom"]["path"])
 
     with aioresponses() as mocked:
@@ -42,10 +41,9 @@ async def test_fetch_result(mocker,
                    status=200,
                    body=json.dumps(success_result_response))
 
-        async with urlscan_async.UrlScanAsync(
-            api_key=test_urlscan_params["api_key"],
-            data_dir=test_urlscan_params["data_dir"]) as urlscan:
-            actual: Dict[str, str] = await urlscan.fetch_result(test_urlscan_params["uuid"])
+        async with urlscan.UrlScan(api_key=test_urlscan_params["api_key"],
+                                   data_dir=test_urlscan_params["data_dir"]) as url_scan:
+            actual: Dict[str, str] = await url_scan.fetch_result(test_urlscan_params["uuid"])
             mock_download_screenshot.assert_called_once_with(test_urlscan_params["screenshot"]["link"])
             mock_download_dom.assert_called_once_with(
                 test_urlscan_params["uuid"],
@@ -61,7 +59,7 @@ async def test_fetch_result(mocker,
 async def test_download_screenshot(mocker,
                                    test_urlscan_params: Dict[str, Any],
                                    screenshot_response: Any) -> None:
-    mock_save_file = mocker.patch("src.urlscanio.urlscan_async.UrlScanAsync.save_file")
+    mock_save_file = mocker.patch("src.urlscanio.urlscan.UrlScan.save_file")
     mock_save_file.return_value = await return_async(None)
 
     with aioresponses() as mocked:
@@ -69,9 +67,9 @@ async def test_download_screenshot(mocker,
                    status=200,
                    body=screenshot_response)
 
-        async with urlscan_async.UrlScanAsync(api_key=test_urlscan_params["api_key"],
-                                              data_dir=test_urlscan_params["data_dir"]) as urlscan:
-            actual: str = await urlscan.download_screenshot(test_urlscan_params["screenshot"]["link"])
+        async with urlscan.UrlScan(api_key=test_urlscan_params["api_key"],
+                                   data_dir=test_urlscan_params["data_dir"]) as url_scan:
+            actual: str = await url_scan.download_screenshot(test_urlscan_params["screenshot"]["link"])
             assert str(test_urlscan_params["screenshot"]["path"]) == actual
             mock_save_file.assert_called_once_with(test_urlscan_params["screenshot"]["path"], screenshot_response)
 
@@ -80,7 +78,7 @@ async def test_download_screenshot(mocker,
 async def test_download_dom(mocker,
                             test_urlscan_params: Dict[str, Any],
                             dom_response: Any) -> None:
-    mock_save_file = mocker.patch("src.urlscanio.urlscan_async.UrlScanAsync.save_file")
+    mock_save_file = mocker.patch("src.urlscanio.urlscan.UrlScan.save_file")
     mock_save_file.return_value = await return_async(None)
 
     with aioresponses() as mocked:
@@ -88,8 +86,8 @@ async def test_download_dom(mocker,
                    status=200,
                    body=dom_response)
 
-        async with urlscan_async.UrlScanAsync(api_key=test_urlscan_params["api_key"],
-                                              data_dir=test_urlscan_params["data_dir"]) as urlscan:
-            actual: str = await urlscan.download_dom(test_urlscan_params["uuid"], test_urlscan_params["dom"]["link"])
+        async with urlscan.UrlScan(api_key=test_urlscan_params["api_key"],
+                                   data_dir=test_urlscan_params["data_dir"]) as url_scan:
+            actual: str = await url_scan.download_dom(test_urlscan_params["uuid"], test_urlscan_params["dom"]["link"])
             assert str(test_urlscan_params["dom"]["path"]) == actual
             mock_save_file.assert_called_once()
